fix all build errors
